+++ design/example_seqdia_test.go
@@ -27,6 +27,6 @@ func ExampleSequenceDiagram() {
 	d.Link(db, srv, "Rows")
 	d.Link(srv, srv, "Transform to view model").Class = "highlight"
 	d.Skip()
-	d.Link(srv, cli, "Send HTML")
+	d.Return(srv, cli, "Send HTML")
 	d.SaveAs("img/app_sequence_diagram.svg")
 }